fix(frontend): Guarantee synced key pair and delegation for `authClient` (#9153)
# Motivation

When multiple OISY tabs are open in the same browser, each creates a new
`authClient` object with its own key pair. Since keys are stored in
IndexedDB (IDB) as a security measure to avoid key injection, the last
loaded tab overwrites the cached keys.

This causes a problem: if a user logs in on a tab that is not the
"latest" one (without refreshing first), the existing `authClient` in
that tab uses a key pair different from the one currently cached in IDB.
It will then request a delegation for those non-cached keys, and store
that delegation in IDB too. As a result, the cache ends up with a
delegation that does not match the cached key.

Later, if we recreate the `authClient` (e.g. inside a worker) or refresh
another tab, the mismatch between keys and delegation leads to invalid
signatures and errors.

To prevent this, we must ensure that after login, the correct key is
always cached in IDB, overwriting any previously stored key. This
guarantees that delegation and keys remain in sync.

# Credits

Big thanks to @ilbertt for debugging and providing a solution! Hopefully
directly in `AuthClient` in the future, eheh!

# Changes

- Use the login callback to get the key pair of the current `authClient`
object (the one used to fetch a delegation).
- Cache such key pair, overwriting any existing one in IDB cache.

# Note

This does not affect the single-async-call restriction of some browsers,
since the callback happens AFTER the login is successful.

# Tests

There are several ways to test. The simplest reproduction is:

1. Open tab 1
2. Open tab 2
3. Login in tab 1 (without refreshing)
4. ERROR!

### Before


https://github.com/user-attachments/assets/da7412fc-81b9-4ff2-81ad-7604426e6627

### After



https://github.com/user-attachments/assets/a0ad866f-f585-408e-afe2-f8ce0c23eb9c

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>