feat(frontend): Always load all Ethereum transactions (#9197)
# Motivation

Some of the sprinkles programs require to have the list of the user
transactions too to work. For Ethereum and EVMs, the transactions
loaders are created only in the Activity page and in the transactions
list for each token.

Now, we need for all the pages (rewards, assets, etc). So, we move the
main loader to the global `Loaders` component. And we remove the
listeners from the transactions lists too.