feat(frontend): Broadcast login success among tabs to resync auth (#9194)
# Motivation

When multiple OISY tabs are open in the same browser, each creates a new
`authClient` object with its own key pair. Since keys are stored in
IndexedDB (IDB) as a security measure to avoid key injection, the last
loaded tab overwrites the cached keys.

This causes a problem: if a user logs in on a tab that is not the
"latest" one (without refreshing first), the existing `authClient` in
that tab uses a key pair different from the one currently cached in IDB.
It will then request a delegation for those non-cached keys, and store
that delegation in IDB too. As a result, the cache ends up with a
delegation that does not match the cached key.

Later, if we recreate the `authClient` (e.g. inside a worker) or refresh
another tab, the mismatch between keys and delegation leads to invalid
signatures and errors.

Part of the issue is being fixed in AuthClient directly, but we already
merged a temporary workaround (PR
https://github.com/dfinity/oisy-wallet/pull/9153).

To try and solve it, we implement a [broadcast
channel](https://developer.mozilla.org/en-US/docs/Web/API/Broadcast_Channel_API)
and force-reload the authclient/identity in any tabs if one tab logs in.

# Credits

Big thanks to @peterpeterparker for the solution!

# Changes

- Create method to force synchronization in the `authStore`.
- Create class `AuthBroadcastChannel` that allows to initialize and
close an OISY broadcast channel, plus it can post the "login successful"
message.
- The handling of the messages will sync the `authStore`, as explained,
but ONLY for same origin and correct message.
- Adapt the main layout to create the channel on mounting and destroying
it on unmounting.
- Trigger the posting of the message after every successful login.

# Tests

Added some unittests, mocking directly the `BroadcastChannel` class.