feat(frontend): Optional toolbar in SendReview component (#9234)
# Motivation

We want to reuse the component `SendReview` in the WalletConnect flow:
this will maintain consistency of style and layout.

However, WalletConnect review requires specific actions on approval and
rejection, while `SendReview` has already set the actions.

In this PR we provide a custom toolbar that otherwise will default to
the existing one.


# Changes

- Adapt the Props of `SendReview` to accept either the actions callbacks
or a completely new toolbar.
- Adapt the code

# Tests

Existing tests should work as expected.