feat(devops): Do not floor coverage thresholds (#9232)
# Motivation

We have reached a stable rhythm for coverage thresholds, so we won’t be
increasing them aggressively with each PR. However, keeping a strict
floor limit introduces the risk of falling close to that minimum and
then needing to add a large batch of new tests all at once.

By removing the fixed floor for threshold values, we can maintain the
right level of coverage while avoiding frequent or sudden updates to the
limits.