fix(frontend): Remove ERC20 dependency from menu buttons (#9266)
# Motivation

The menu buttons were all enable-dependent on the ERC20 tokens store
being initialized, however it is not coherent, since they are used for
all networks and none of them really has any direct dependency on ERC20
tokens.

Furthermore, it means that if the Ethereum/EVM networks are disabled,
all these components would be disabled too, and it does not really work
for the user.