feat(frontend): disable to install oisy as PWA (#9103)
# Motivation

When OISY is installed as a PWA standalone it is not possible to connect
other DEX's that are running in PWA standalone mode to OISY. To resolve
this issue we have changed the `display-mode` setting to `browser`,
which prevents OISY from being installed as a PWA anymore.
Users who have already installed OISY as a PWA are still affected. For
them, we display a warning message to inform them that they should
reinstall it.

# Changes

- activate browser mode

# Tests

**New:**
<img width="869" height="376" alt="image"
src="https://github.com/user-attachments/assets/f1a8e924-94f3-4b4d-a468-537d09d00540"
/>

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>