feat(frontend): Remove ERC20 dependency from the manage tokens button (#9303)
# Motivation

The manage tokens button was enable-dependent on the ERC20 tokens store
being initialized, however it is not coherent, since it is used for all
networks and it has no direct dependency on ERC20 tokens.

Furthermore, it means that if the Ethereum/EVM networks are disabled, it
would be disabled too, and it does not really work for the user.