feat(frontend): handle ck tokens in parseShowBalanceToolArguments (#9315)
# Motivation

We want to include ck tokens into the equation when parsing the
show_balance tool response. The only caveat is to make sure it's not the
first element of the array as we use such an item as a header for the
whole group. To prevent it, we always add ck twin tokens at the end of
the array.

Note: it doesn't need to work the other way around, hence when user asks
to show e.g. ckBTC balance we do ignore BTC.

<img width="404" height="1160" alt="Screenshot 2025-10-01 at 15 21 20"
src="https://github.com/user-attachments/assets/57328d44-081d-46a4-9a36-a0ac47132d70"
/>