feat(frontend): Remove ERC20 dependency from the send tokens list (#9305)
# Motivation

The send tokens list sets the loading state based on the ERC20 tokens
store being initialized, however it is not coherent, since it is used
for all networks and it has no direct dependency on ERC20 tokens.

Furthermore, it means that if the Ethereum/EVM networks are disabled, we
would be waiting for the store to load infinitely.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>