fix(frontend): Remove concurrency from `Banner` component (#9330)
# Motivation

In the `Banner` component, the two separate banners (one for the
environment and one for PWA) are using the same variable to be visible
or not. This causes each banner to close both, and that is not intended
by the user.

To fix it, we assign separate variables to each closing action.