fix(frontend): Remove concurrency of accepting agreements (#9347)
# Motivation

When mounting the component `AgreementsGuard`, the user is
auto-accepting the agreements.

Since it depends on derived stores of the user profile, it could be
called several times consecutively. It is not an issue, but we can spare
calls.

The issue is that accept `acceptAgreements` is asynchronous, but the
rune `$effect` does not accept `async` function, so it will not await
`acceptAgreements` to finish. So, even using a flag to avoid duplicate
calls is not that easy: the flag could be set back to `false` BEFORE
`acceptAgreements` is finished.

However, we have a very useful service called `queryAndUpdate` that
executes canisters calls (query and/or update) and has callbacks in case
of success or failure.


# Changes

- Use `queryAndUpdate` to run `acceptAgreements`:
  - only update call;
  - set the accepting flag to false in case both of success or error;
- avoid running `acceptAgreements` if it is already running (`accepting
=== true`).
- For good measure, we avoid running it completely if there are no
agreements to accept.

# Tests

Created a new test that was failing before this PR:

<img width="1287" height="944" alt="Screenshot 2025-10-02 at 14 23 55"
src="https://github.com/user-attachments/assets/24efb74d-04e1-4926-81bf-e45be0f9be68"
/>

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>