feat(frontend): Util to concatenate the agreements links (#9349)
# Motivation

We will need to display a banner for the updated agreements for the
users.

That entails that we will need to be able to concatenate 3, 2, or 1
agreements in the same sentence, with their respective links. For
example:

"We have updated our Policy Privacy and License Agreement".

There is a tweak: in other languages we have the gender and the plural
too. For example in Italian:

"Abbiamo aggiornato la nostra Politica di Privacy e i nostri Termini
d'Uso"

So, we need to consider a concatenation with articles too.

# Changes

- Add determinative articles to the agreements (and possessive too).
- Create util to map each agreement to the link-component and the
correct enunciation.
- Use `Intl` to concatenate based on the current language.

# Note

The new functions in should have gone in the `agreements.utils.ts`
module. However, in their usage there is a circular reference somewhere
in the build that makes it fail. I was still unable to find the source
of the issue, so I decided to keep them in a separate module for now. To
be investigated.

# Tests

Added tests cases in English, Italian and German for good measure.