feat(frontend): Display approve spender (#9320)
# Motivation

Based on #9280.

We want to display the spender account for Icp/Icrc transactions.

# Changes

- Refactored mappings for transactions
- Adjusted fee and amount definitions so it is consistent across the
different transaction types
- Display spender for approve transactions in modal and transaction card

# Tests

All new additions should be tested fully:

- Transaction cards tested by #9346 which ensures that transactions of
the other networks are not affected by the changes
- Displaying spender in contact cards tested by tests in #9342
- Added tests for ensuring transaction card label is correct including
allowance for approve
- Extended tests to ensure contacts are displayed in transaction cards
and that it shows the aprove spender contact for approve transactions

Manually tested as well on FE3

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>