feat(frontend): integrating PowProtector component (#9302)
# Motivation

This PR finalizes the PoW protection system by integrating the
PowProtector component.
The complete feature, including all related changes, is tracked in PR
#7409.

# Changes

**AddressGuard.svelte:**
Updated to handle initialization only if the PoW feature is disabled.

**Loaders.svelte:**
Integrated the new PowProtector component into the loader hierarchy,
positioned immediately after the user profile is loaded. If the Pow
feature ist enabled this call must be disabled because it triggers an
allowance call that requires a nonce, which is now managed by the PoW
allowance process.


**loader.services.ts:**
Updated to handle initialization only if the PoW feature is disabled. If
the Pow feature ist enabled this call must be disabled because it
triggers an allowance call that requires a nonce, which is now managed
by the PoW allowance process.

# Tests

Tested on fe2 with VITE_POW_FEATURE_ENABLED=false (which is the default
value)