feat(frontend): Specific component for values in WalletConnect modals (#9370)
# Motivation

As per decision, the WalletCOnnect modal should not give the impression
that the transaction is native of OISY, meaning, as hint, that it should
have a different style from the normal modals. We are opting for a
simple one.

So, as first step we extract a common component for the component
`WalletConnectModalValue`.

Example:

<img width="424" height="79" alt="Screenshot 2025-10-03 at 17 12 24"
src="https://github.com/user-attachments/assets/7e0b4901-8ebc-45e7-8b91-f7183a396adc"
/>