fix(frontend): Correct exchange rate for Solana fees (#9427)
# Motivation

The Solana fee were not using the exchange rate of the Solana native
token they refer to, but the ones of the token that was going to be
sent.

# Changes

- Add exchange rate store in the Fee context of Solana.
- Set the store correctly in the same place as the rest of the context.
- Use the store to calculate the FIAT amount of the fees.

# Tests

Adapted tests and a practical one too: in local we don't have exchange
rate for USDC Devnet, and it was using that to calculate (hence
`undefined`).

### Before

<img width="682" height="784" alt="Screenshot 2025-10-06 at 15 55 09"
src="https://github.com/user-attachments/assets/561eea03-8db6-4d23-bd5a-bafd512febf1"
/>

### After

<img width="623" height="761" alt="Screenshot 2025-10-06 at 15 44 06"
src="https://github.com/user-attachments/assets/8138dcda-dbc0-468b-8df5-ed511ce3c243"
/>