feat(frontend): added new swap icrc tokens store (#9439)
# Motivation

We need to remove the dependency on KongSwap tokens and use
allIcrcTokens as a fallback, but with an additional property isIcrc2,
derived from icrc1_supportedStandard.
The scope of this PR includes creating a new store that will contain all
of these tokens with the added property.

# Changes

Created a new swap-icrc-tokens.store.

# Tests

Added test coverage for the new store.