fix(frontend): Do not show expired toast for locked state (#9452)
# Motivation

If the user has more than one tab open with OISY, when it locks one, the
other will reload the page BUT show a toast message for the expired
session. And that is not correct.

The reason being that the worker for idle sign out does not
differentiate between locked state and expired session.

# Changes

- Expose locked key for local storage.
- When calling the idle sign out, check first if locked or not, and
appropriately return the correct message.

# Tests

Added some tests and a practical one:

### Before


https://github.com/user-attachments/assets/d095adfc-f74b-4196-a6ea-9e5d0f97fc9d

### After



https://github.com/user-attachments/assets/f6e28666-834f-4d18-a53a-bd89c1dd7f2b