feat(frontend): Remove force sign-out during identity checks (#9487)
# Motivation

In several places in the code, we check if the user is still signed in
or not or if identity is nullish, mostly for the usage of a non-nullish
identity.

In most cases, as safe-measure, we execute the `nullishSignOut` service
that forces a sign-out with a warning. However, there is already a
worker that acts like that, so it is theoretically unnecessary.