refactor(frontend): Rename Nft consent component (#9531)
# Motivation

We rename this component so it acts as a general guard for displaying
Nfts, instead of only for media consent. For that we just rename it in
this PR, and later extend it to display placeholders in case of
erroneous image data.

# Changes

Renamed all occurrences and the component itself

# Tests

Renamed all tests