feat(build): Copy generated bindings in subdirectory of `icp-bindgen` (#9547)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/9521, we are going to
introduce the use of library
[`@icp-sdk/bindgen`](https://js.icp.build/bindgen/latest/) that helps us
creating the bindings, substituting effectively `dfx generate`.

However, since the new library saves the generated files in a
sub-directory /declarations, we need to adapt our code. And the change
is quite through a lot of files.

To easy this transition, we first copy the current bindings in the same
directory, and subsequently we will adapt the code, one by one. Finally,
we will switch to `icp-bindgen`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>