feat(frontend): Show IC token metadata error only for Custom tokens (#9556)
# Motivation

When we initially load the default tokens, we show an error if the
ledger canister fails to give us the metadata.

However, it does not really make sense to do it for the default tokens,
but only for the custom tokens: the user is interested in the error ONLY
if the token was among the enabled ones.

# Changes

- Remove the toast message from the loading of default tokens.
- Add the same message to the loading of custom tokens.
- Show the error message after the update call.

# Tests

Added tests.