refactor(frontend): Refer to new path for `icp_swap_factory` bindings (#9566)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/9521, we are going to
introduce the use of library
[`@icp-sdk/bindgen`](https://js.icp.build/bindgen/latest/) that helps us
creating the bindings, substituting effectively `dfx generate`.

However, since the new library saves the generated files in a
sub-directory /declarations, we need to adapt our code. And the change
is quite through a lot of files.

To easy this transition, in this PR we just change the imports that
refers to `icp_swap_factory` bindings to the new path.