feat(frontend): Move QR `onClick` event in `ReceiveActions` (#9631)
# Motivation

The `onClick` event for component `ReceiveActions` is strictly linked to
the QR action (the button that should show the QR code).

So, for manageability, we can add this property to the type
`ReceiveQRCodeAction` directly: it will be anyway given ONLY when the QR
actions are enabled.

# Changes

- Add `onClick` to the type `ReceiveQRCodeAction` when they are enabled.
- Move the `on:click` events inside the prop `qrCodeAction` for all
consumers.

# Tests

Current tests work as expected.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>