fix(frontend): Avoid repeated tokens when managing enabled and disabled (#9685)
# Motivation

When we manage tokens, we are currently indexing them with a string that
defines the token ID and the network ID, both `symbols` but parsed as
`string`.

However, it may happen to have tokens with same network and same ID (for
example some NFTs).

We can directly use instead the `TokenId` type that is a `symbol`: each
`symbol`, even if it has the same `description`/`string`, is treated as
a different unique object. So, using it for indexing guarantees no
problems in confusing one with the other.

# Changes

- Change the `Record` of modified tokens to be a `Map` in `ManageTokens`
component.
- Adapt the flow to add/remove/update a modified token.

# Tests

### Before


https://github.com/user-attachments/assets/0af388ee-188c-4dca-a82d-b4e0dd4c2ee9


### After



https://github.com/user-attachments/assets/510559f8-09bd-48e0-860a-eee4ad5dbfb7