fix(frontend): Nft Hide should only disable media if allowExternalContentSource is undefined (#9692)
# Motivation

Currently, marking a collection as spam always disables external media.
We changed the hide to behave the same, but it should only disable
external media if the user has not yet actively chosen to enable or
disable media.

# Changes

Updated condition of when to set allowExternalContentSource to false

# Tests

Added test to verify this behavior