feat(frontend): fixed velora delta swap destAmount (#9731)
# Motivation

We need to fix Velora delta swap receive amount

# Changes

Start using new field from velora response with fallback.

We need to use `destAmountAfterBridge` in `bridgeInfo` for delta swaps,
if available, since this value is not mentioned in the documentation.
However, after discussions with the Velora team, we confirmed that this
value is already formatted correctly with the required token amount
precision.

# Tests

Adjusted tests