fix(frontend): Bad z-index in `StickyHeader` (#9818)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/9765 I forgot to push
the commit that fixes the z-index of `StickyHeader`.

NOTE: `z-3` does not exists among the Tailwind classes, but that was the
value that was assigned before the refactoring, so I would prefer to
keep it as it was. We can revisit in another PR.