feat(lint): Temporary disable Svelte rule `no-navigation-without-resolve` (#9873)
# Motivation

There is a new rule in the `recommended` suite of Svelte Lint plugin:
[no-navigation-without-resolve](https://sveltejs.github.io/eslint-plugin-svelte/rules/no-navigation-without-resolve/).

The rule enforces resolving the internal URLs. However, in our code it
raises quite a lot of warnings (see CI
https://github.com/dfinity/oisy-wallet/actions/runs/18709289138/job/53353612091?pr=9761).

We have an higher priority to bump the OISY signer, so for now we simply
disable the rule, and work on the fixes step-by-step.