feat(frontend): Method to fetch NFT metadata in `AlchemyProvider` (#9848)
# Motivation

When we are showing the transactions list, if there are NFT that are
sent out, we will not have them anymore in our store and consequently
the images will not appear in the single transaction.

To avoid it, we could try to fetch the NFT metadata on-demand, when this
happens.

As first step we need to expose the method `getNftMetadata` of the
Alchemy provider (see //
https://www.alchemy.com/docs/reference/nft-api-endpoints/nft-api-endpoints/nft-metadata-endpoints/get-nft-metadata-v-3).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>