fix(frontend): Use correct map when modifying tokens in Assets page (#9855)
# Motivation

When we manage tokens in the Assets page through the filter, we are
currently indexing them with the `TokenId` that is a `symbol` object.
However it is quite complicated to extract the values from a `Record`
indexed with symbols.

So, we can use a SvelteMap and simply using its built-in methods.