feat(frontend): Nft assets page routing (#9889)
# Motivation

We add the "routing" by query params to the assets page. This will
currently have no effect but once the routing gets switched over, these
act as the collection and nft pages.

# Changes

Added routing conditions and components

# Tests

Since the pages rely on the routing params, the functionality of the
routing is already covered heavily. Also, this doesnt change the routing
to query params yet, so we need to adjust the tests once the routing has
switched to query params.