feat(frontend): Move PWA Banner to sign page of OISY (#9877)
# Motivation

Credits to @sbpublic :

We don't really need the PWA warning banner in the entire OISY, in case
of a PWA. The real issue with PWA is born with missing connections
during the OISY signer flow. So, we can show it ONLY in the `sign`
routes.


<img width="567" height="542" alt="Screenshot 2025-10-22 at 14 34 40"
src="https://github.com/user-attachments/assets/13b7266b-5464-4844-84ed-c8ed4acd1a74"
/>