chore(frontend): Make service `initPlausibleAnalytics` fake-async for state of the art (#9904)
# Motivation

Service `initPlausibleAnalytics` is used during initialization together
with other async functions:

```ts
await Promise.allSettled([syncAuthStore(), initPlausibleAnalytics(), i18n.init()]);
```

However, it is not an asynchronous function, so the statement is a bit
incongruous. To have a state-of-the-art code we forcefully make it
`async`.

# Note on Lint

I prefer to make an exception to the lint rule `require-await` instead
of returning a `Promise.resolve`: if and when `initPlausibleAnalytics`
will become really `async`, the linter will warn us of an unused
exception. It seems cleaner.