feat(frontend): Remove unnecessary initial progress loader modal (#9984)
# Motivation

The initial modal loader is currently used to give a feedback to the
user especially when the addresses are not derived in the frontend.

That is currently happening ONLY in the local instance (DEV usage). And
in any case, the entire UI is made (or should be made) to handle the
loading of the addresses with skeletons.