feat(frontend): REVERT: add interval in `LoaderNfts` (#9988)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/9827, I made the
`LoaderNfts` more reactive. And I removed the interval, expecting all
cases to be covered by the reactivity.

However, if a collection is already enabled and we receive a new NFT for
such collection, the list is not updated. So, to avoid this issue, we
create an interval service to fetch them every time.