feat(frontend): Stay on correct assets tab when navigating (#9974)
# Motivation

When we navigate away from the assets page (tokens or Nfts), it
currently always gets back to tokens when we return to the assets page.
We want to return to the Nfts tab for example if that has been visited
last.

# Changes

- Added condition to the main nav item

# Tests

Created basic tests for the main nav items component and specific ones
for this feature

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>