fix(frontend): Incorporate network only if user selected (#10015)
# Motivation

Currently, when an Nft or collection page is opened, the network of that
collection is added to the nav items, even if the user didnt explicitely
selected it via network switcher.

We want the query param only when the user selected it.

# Changes

Use userSelectedNetwork to build the Urls

# Tests

Added tests

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>