feat(frontend): Add section to collection (#10060)
# Motivation

To easier filter out hidden and spam Nft collections we add it to the
collection object.

Since we refresh the data after changing the section (#9734) this
mapping is called after any change to the section. This means we are
safe to read from the collection as it should always reflect the newest
state.

# Changes

- Added section to collection schema
- Added section to collection mapping

# Tests

Added tests

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>