feat(frontend): Always enable new collections (#10100)
# Motivation

Until now we were enabling automatically new collections ONLY when
Alchemy does not define them as Spam. However, we checked and it does
not seem we can really really on this information (some collections are
wrongly defined as spams).

So, for now we enable them all if they are new.