fix(frontend): Refresh NFTs by network (#10104)
# Motivation

In `LoaderNFts` component, when we refresh the NFTs per collection, we
do it by network, adding it to the existing store (removing duplicates
too).

However, if we send out an NFT, when we refresh, we will not receive it
from the provider, so, we will not remove it from the store.

To fix this, we can simply add the NFTs by network, completely removing
the ones already saved for that specific network.

# Changes

- Add new method `setAllByNetwork` in NFT store.
- Use the new method in `LoaderNfts`.

# Tests

Added tests and a practical one:



https://github.com/user-attachments/assets/060de98a-8517-4380-ad55-a4c0f584d4e3

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>