feat(frontend): Wrap OISY logo in the seasonal guard (#10162)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/10059, we added the
Halloween logo temporarily to OISY. However, it would be useful to have
it automatically triggered in specific periods.

So, we will wrap the OISY logo in the SeasonalGuard component, and
revert the before-halloween assets.

# Changes

- Restore the pre-halloween assets.
- Create component for seasonal OISY logo, mixing current existing
Halloween logo and "normal" one.