feat(frontend): Send user snapshots for NFT (#10158)
# Motivation

We want to start sending snapshots of the NFTs too. However, we still
need to adapt a bit the Reward Canister (RC).

For example:

- `Standard`: we cannot tell the RC the difference between NFT and
Fungible Token currently. So, for now we create a custom string for the
NFT.
- `Balance`: we should be able to send separate NFT for each collection
with their specific balance, but for now we just send the number of
different NFT per collection as balance.

# Changes

- Map the token symbol of NFT as a string with separators and necessary
data.
- Map the token balance of NFT with the number of difference NFT ID
(different NFTs).

# Tests

Added tests