feat(frontend): Load ETH transactions from cache reactively (#10171)
# Motivation

We are loading the ETH transactions from the cache only on mounting of
components, so we could risk that the list of tokens is not really set.

Since we want more reactivity, we would like to trigger the cache
loading everytime the list is updated.

NOTE: In any case, nothing will be saved from cache if the transaction
store already has entries for that specific tokens.

# Changes

- Trigger the ETH transaction loader when the enable tokens list
changes.

# Tests

Current tests should work as expected.