feat(frontend): Prioritize tokens with empty transactions in `LoaderMultipleEthTransactions` (#10178)
# Motivation

In component `LoaderMultipleEthTransactions`, even if it had a bit of
complexity, we prefer to prioritise the tokens that have empty
transaction store, because they are more likely the ones that are still
not loaded.

The idea is to visually be "faster" to load the transactions in case of
a lot of tokens.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>