feat(frontend): migrate agent-js to icp-sdk (#9192)
# Motivation

Migrate AgentJS `@dfinity/...` imports to `@icp-sdk/core` and
`@icp-sdk/auth`. Those multiple entries libraries are currently using
the AgentJS libraries we are using underneath, therefore no functional
changes are to be expected.

# Notes

If we miss an import in this PR it isn't an issue. `@icp-sdk/core` is
basically a proxy for `@dfinity/*`, therefore those libraries are still
available in the repo.

# Changes

- Replace `@dfinity/*` libraries by `@icp-sdk/core` and `@icp-sdk/auth`
- Update `npm run update:agent` to update those libraries in the future
- Update related imports to used the new sub-entries path (search and
replace)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>